[refactor] 공통 클래스 사용을 위해 패키지 구조 변경
