+++ src/main/java/project/calculator/NumberType.java
@@ -1,4 +1,4 @@
-package calculator;
+package project.calculator;
 
 public enum NumberType {
     INTEGER, DECIMAL
