+++ src/main/java/project/game/Car.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 public class Car {
