+++ src/main/java/project/game/MoveRule.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 public interface MoveRule {
     boolean canMove();
