+++ src/main/java/project/game/RacingGame.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 import java.util.ArrayList;
 import java.util.List;
