+++ src/main/java/project/game/RandomNumberProvider.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 public interface RandomNumberProvider {
     int getRandomNumber();
