+++ src/main/java/project/game/RandomNumberProviderImpl.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 import java.util.Random;
