+++ src/test/java/project/calculator/CalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package project.calculator;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
