+++ src/test/java/project/calculator/ExpressionTest.java
@@ -1,9 +1,7 @@
-package calculator;
+package project.calculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.*;
