+++ src/test/java/project/game/CarTest.java
@@ -1,4 +1,4 @@
-package game;
+package project.game;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
